Code cleanup and fixes

+++ src/main/java/com/jelly/farmhelper/features/Autosell.java
@@ -93,7 +93,7 @@ public class Autosell {
             if (mc.thePlayer.inventory.getFirstEmptyStack() == -1 && !fullCheck) {
                 LogUtils.debugLog("[AutoSell] Started inventory full watch");
                 fullCheck = true;
-                checkTimer.schedule(TimeUnit.SECONDS.toMillis((long) FarmHelper.config.inventoryFullTime));
+                checkTimer.schedule(TimeUnit.SECONDS.toMillis(FarmHelper.config.inventoryFullTime));
                 fullCount = 1;
                 totalCount = 1;
             } else if (fullCheck && !checkTimer.passed()) {
